backend/pyproject.toml
[tool.poetry]
name = "cern-academic-training"
version = "0.1.0"
description = ""
authors = ["Harris Tzovanakis <me@drjova.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
django-filter = "^21.1"
django-elasticsearch-dsl = "^7.2.2"
django-elasticsearch-dsl-drf = "^0.22.2"
dj-database-url = "^0.5.0"
psycopg2 = "^2.9.3"
sentry-sdk = "^1.5.4"
gunicorn = "^20.1.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
